test: try to stabilize thumbnails tests, extract common helpers
